+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Semgrep analyzer changelog
 
+## v4.1.0
+- upgrade `gitlab.com/gitlab-org/security-products/analyzers/ruleset` version [`v1.4.1` => [`v2.0.0`](https://gitlab.com/gitlab-org/security-products/analyzers/command/-/releases/v2.0.0)] (!245)
+- enable support for loading remote ruleset configs (!245)
+
 ## v4.0.1
   - Update `eslint.yml` rules to [sast-rules](https://gitlab.com/gitlab-org/security-products/sast-rules) version `1.3.0` (!244)
     - Remove security prefix [sast-rules!94](https://gitlab.com/gitlab-org/security-products/sast-rules/-/merge_requests/94)